docs: add LineChartConnectNulls example in TypeScript (.tsx)
