+++ src/docs/exampleComponents/LineChart/LineChartConnectNulls.tsx
@@ -1,0 +1,57 @@
+import { CartesianGrid, Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
+
+const data = [
+  { name: 'Page A', uv: 4000 },
+  { name: 'Page B', uv: 3000 },
+  { name: 'Page C', uv: 2000 },
+  { name: 'Page D' },
+  { name: 'Page E', uv: 1890 },
+  { name: 'Page F', uv: 2390 },
+  { name: 'Page G', uv: 3490 },
+];
+
+export default function Example() {
+  return (
+    <div style={{ width: '100%' }}>
+      <ResponsiveContainer width="100%" height={200}>
+        <LineChart
+          width={500}
+          height={200}
+          data={data}
+          margin={{
+            top: 10,
+            right: 30,
+            left: 0,
+            bottom: 0,
+          }}
+        >
+          <CartesianGrid strokeDasharray="3 3" />
+          <XAxis dataKey="name" />
+          <YAxis />
+          <Tooltip />
+          <Line type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
+        </LineChart>
+      </ResponsiveContainer>
+
+      <ResponsiveContainer width="100%" height={200}>
+        <LineChart
+          width={500}
+          height={200}
+          data={data}
+          margin={{
+            top: 10,
+            right: 30,
+            left: 0,
+            bottom: 0,
+          }}
+        >
+          <CartesianGrid strokeDasharray="3 3" />
+          <XAxis dataKey="name" />
+          <YAxis />
+          <Tooltip />
+          <Line connectNulls type="monotone" dataKey="uv" stroke="#8884d8" fill="#8884d8" />
+        </LineChart>
+      </ResponsiveContainer>
+    </div>
+  );
+}
